service/stripe-forward-integration/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
confluent-kafka==2.3.0
idna==3.6
python-dotenv==1.0.1
requests==2.31.0
stripe==8.8.0
typing_extensions==4.10.0
urllib3==2.2.1
psycopg2-binary